requirements-dashboard.txt
netron
onnx
streamlit
tensorflow-macos
tf2onnx
torch
